move config to separate file, enable debugging
