+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "test": "mocha --exit",
     "build": "npm install",
-    "start": "node index.js"
+    "start": "node index.js",
+    "debug": "node --nolazy --inspect-brk=9229 index.js"
   },
   "author": "",
   "license": "ISC",
