+++ workspace.code-workspace
@@ -1,0 +1,12 @@
+{
+	"folders": [
+		{
+			"path": "."
+		}
+	],
+	"settings": {
+		"appService.advancedCreation": true,
+		"appService.zipIgnorePattern": ["node_modules/**"],
+		"appService.showBuildDuringDeployPrompt": true
+	}
+}
